// src/companies/enums/comapanies.enum.ts
export enum GetAllTabEnum {
  PROSPECT = 'prospect',
  DISSOLUTION = 'dissolution',
  CHATBOT = 'chatbot',
}

export enum CompanyStatusEnum {
  RECENTLY_DECIDED_SHUTDOWN = 'RECENTLY_DECIDED_SHUTDOWN',
  MIDDLE_OF_SHUTDOWN = 'MIDDLE_OF_SHUTDOWN',
  PARTIAL_UNWIND = 'PARTIAL_UNWIND',
  DISTRESSED_SALE = 'DISTRESSED_SALE',
  UNDECIDED = 'UNDECIDED',
}

export enum CompanyStageEnum {
  //prospect client
  NEW_REGISTERED_CLIENT = 'NEW_REGISTERED_CLIENT',
  PENDING_INTAKE_CALL = 'PENDING_INTAKE_CALL',
  CONTRACT_AND_FEE_PENDING = 'CONTRACT_AND_FEE_PENDING',
  //disolution client
  DECISIONS_AND_APPROVALS = 'DECISIONS_AND_APPROVALS',
  ENGAGE_WITH_ENCORE_AND_IP_TRANSFER = 'ENGAGE_WITH_ENCORE_AND_IP_TRANSFER',
  FILING_AND_NOTIFICATIONS = 'FILING_AND_NOTIFICATIONS',
  PUBLIC_ANNOUNCEMENTS_AND_COMMUNICATIONS = 'PUBLIC_ANNOUNCEMENTS_AND_COMMUNICATIONS',
  IP_ASSET_DISPOSITION = 'IP_ASSET_DISPOSITION',
  SETTLEMENTS_AND_CLOSURES = 'SETTLEMENTS_AND_CLOSURES',
  REGULATORY_COMPLIANCE_AND_RECORD_KEEPING = 'REGULATORY_COMPLIANCE_AND_RECORD_KEEPING',
  //chatbot
  AI_CHATBOT_ONLY = 'AI_CHATBOT_ONLY',
}

export enum CompanyServicesEnum {
  DISSOLUTION = 'DISSOLUTION',
  AI_CHATBOT = 'AI_CHATBOT',
}

export enum CompanyIntellectualPropertyTypeEnum {
  PATENTS = 'PATENTS',
  TRADEMARKS = 'TRADEMARKS',
  COPYRIGHTS = 'COPYRIGHTS',
  TRADE_SECRETS = 'TRADE_SECRETS',
}

export enum CompanyStructureEnum {
  CORPORATION = 'CORPORATION',
  LLC = 'LLC',
  OTHER = 'OTHER',
}